requirements.txt
linkedin-api==2.0.0a5
python-dotenv==1.0.0
schedule==1.2.0
SQLAlchemy==2.0.23
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
python-jose==3.3.0
requests==2.31.0
aiohttp==3.9.1
pytest==7.4.3
black==23.11.0
isort==5.12.0
pytz==2023.3 